split test so it can run in parallel
